58 - Create Cases table & class, updated other pages, & cases page field namespace. Refactor created by & updated by id int to long.

+++ src/main/java/com/jamersc/springboot/financialhub/model/Check.java
@@ -44,14 +44,14 @@ public class Check {
     private Date checkDate;
 
     @Column(name = "created_by")
-    private int createdBy;
+    private Long createdBy;
 
     @Column(name = "created_at", updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
 
     @Column(name = "updated_by")
-    private int updatedBy;
+    private Long updatedBy;
 
     @Column(name = "updated_at")
     @Temporal(TemporalType.TIMESTAMP)
