+++ src/main/java/com/jamersc/springboot/financialhub/model/Fund.java
@@ -31,14 +31,14 @@ public class Fund {
     private Double fundBalance;
 
     @Column(name = "created_by")
-    private int createdBy;
+    private Long createdBy;
 
     @Column(name = "created_at", updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
 
     @Column(name = "updated_by")
-    private int updatedBy;
+    private Long updatedBy;
 
     @Column(name = "updated_at")
     @Temporal(TemporalType.TIMESTAMP)
